Create Inter Quartile method
IQR Rule for Outliers: 
1.	Arrange the data in order.
2.	Calculate the first Quartile(Q1), Third Quartile (Q3) and the inter quartile range (IQR = Q3-Q1)
3.	Compute Q1-1.5*IQR and Q3+1.5*IQR
4.	Anything outside this range is an outlier.
For Example: 
Find the outliers for the following data set :3, 10, 14, 22, 19, 29, 70, 49, 36,   32.
Solution: 
1.	 Find the IQR, Q1(25th percentile) and Q3(75th percentile). 
IQR = 22
Q1 = 14
Q3 = 36
2.	Multiply the IQR you found in Step 1 by 1.5:
IQR * 1.5 = 22 * 1.5 = 33.
3.	Add the amount you found in Step 2 to Q3 from Step 1:
33 + 36 = 69.
This is your upper limit. Set this number aside for a moment.
4.	Subtract the amount you found in Step 2 from Q1 from Step 1:
14 – 33 = -19.
This is your lower limit. Set this number aside for a moment.
5.	Put the numbers from your data set in order:
3, 10, 14, 19, 22, 29, 32, 36, 49, 70

6.	Insert your low and high values into your data set, in order:
-19, 3, 10, 14, 19, 22, 29, 32, 36, 49, 69, 70
7.	Highlight any number below or above the numbers you inserted in Step 6:
-19, 3, 10, 14, 19, 22, 29, 32, 36, 49, 69, 70


How to Find Outliers with the The Tukey Method?
The Tukey method for finding outliers uses the interquartile range to filter out very large or very small numbers.