Add files via upload
Dixon’s Q Test 
Dixon’s Q test, or just the “Q Test” is a way to find outliers in very small, normally distributed, data sets. Small data sets are usually defined as somewhere between 3 and 7 items.
Dixon came up with many different equations to find true outliers.
The most commonly used one is called the R10 or simply the “Q” version, which is used to test if one single value is an outlier in a sample size of between 3 and 7.
Example: Is 167 an outlier in this set of data? Test at the 95% confidence Level (i.e. at an alpha level of 5%).167, 180, 188, 177, 181, 185, 189
Step 1: Sort your data into ascending order. 
167, 177, 180, 181, 185, 188, 189.
	
Step 2: Find the Q statistic using the following formula:
 
Where,
x1 is the smallest (suspect) value,
x2 is the second smallest value,
and xn is the largest value.
Inserting the values into the formula, we get:
Q = (177 – 167) / 189 – 167 = 10/22 = 0.455.

Step 3: Find the Q critical value in the Q table. F